# Improve performance/memory overhead for Spatial Join