--- a/lib/trino-geospatial-toolkit/src/main/java/io/trino/geospatial/GeometryUtils.java
+++ b/lib/trino-geospatial-toolkit/src/main/java/io/trino/geospatial/GeometryUtils.java
@@ -119,6 +119,38 @@ public final class GeometryUtils
         }
     }
 
+    /**
+     * Get the bounding box for an OGCGeometry.
+     * <p>
+     * If the geometry is empty, return a Retangle with NaN coordinates.
+     *
+     * @param ogcGeometry
+     * @return Rectangle bounding box
+     */
+    public static Rectangle getExtent(OGCGeometry ogcGeometry)
+    {
+        return getExtent(ogcGeometry, 0.0);
+    }
+
+    /**
+     * Get the bounding box for an OGCGeometry, inflated by radius.
+     * <p>
+     * If the geometry is empty, return a Retangle with NaN coordinates.
+     *
+     * @param ogcGeometry
+     * @return Rectangle bounding box
+     */
+    public static Rectangle getExtent(OGCGeometry ogcGeometry, double radius)
+    {
+        com.esri.core.geometry.Envelope envelope = getEnvelope(ogcGeometry);
+
+        return new Rectangle(
+                envelope.getXMin() - radius,
+                envelope.getYMin() - radius,
+                envelope.getXMax() + radius,
+                envelope.getYMax() + radius);
+    }
+
     public static boolean disjoint(Envelope envelope, OGCGeometry ogcGeometry)
     {
         GeometryCursor cursor = ogcGeometry.getEsriGeometryCursor();