… core/trino-main/pom.xml
… lib/trino-geospatial-toolkit/pom.xml
… lib/trino-geospatial-toolkit/src/main/java/io/trino/geospatial/KdbTree.java
… lib/trino-geospatial-toolkit/src/main/java/io/trino/geospatial/rtree/Flatbush.java
… lib/trino-geospatial-toolkit/src/main/java/io/trino/geospatial/rtree/HilbertIndex.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/TestGeometryUtils.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/TestKdbTree.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/rtree/BenchmarkFlatbushBuild.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/rtree/BenchmarkFlatbushQuery.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/rtree/BenchmarkJtsStrTreeBuild.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/rtree/BenchmarkJtsStrTreeQuery.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/rtree/RtreeTestUtils.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/rtree/TestFlatbush.java
… lib/trino-geospatial-toolkit/src/test/java/io/trino/geospatial/rtree/TestHilbertIndex.java
… plugin/trino-geospatial/src/main/java/io/trino/plugin/geospatial/SpatialPartitioningInternalAggregateFunction.java
… plugin/trino-geospatial/src/main/java/io/trino/plugin/geospatial/SpatialPartitioningStateFactory.java
… plugin/trino-geospatial/src/test/java/io/trino/plugin/geospatial/TestGeoFunctions.java
… plugin/trino-geospatial/src/test/java/io/trino/plugin/geospatial/TestSpatialJoinOperator.java
… plugin/trino-geospatial/src/test/java/io/trino/plugin/geospatial/TestSpatialPartitioningInternalAggregation.java