Improve performance/memory overhead for Spatial Join.

### Comparable changes:
- Major Improvement
  - Use Flatbush RTree for Spatial Joins (prestodb/presto#13079)
- Bug fixes
  - prestodb/presto#13251
  - prestodb/presto#13403
  - prestodb/presto#14485
- Refactoring
  - prestodb/presto#14545
  - prestodb/presto#14856

This change enhance reliability for Spatial Join. 
It reduces the amount of memory for building spatial join index significantly.

According to the local test scenario, the peak memory reduced 96%.

According to the benchmark tests, the performance also improved like below,

# Original
Benchmark                           (numBuildEnvelopes)  Mode  Cnt     Score    Error  Units
BenchmarkJtsStrTreeQuery.rtreeQuery                1000  avgt   20     0.594 ±  0.006  us/op
BenchmarkJtsStrTreeQuery.rtreeQuery                3000  avgt   20     1.062 ±  0.102  us/op
BenchmarkJtsStrTreeQuery.rtreeQuery               10000  avgt   20     2.434 ±  0.106  us/op
BenchmarkJtsStrTreeQuery.rtreeQuery               30000  avgt   20     6.496 ±  0.055  us/op
BenchmarkJtsStrTreeQuery.rtreeQuery              100000  avgt   20    21.323 ±  1.238  us/op
BenchmarkJtsStrTreeQuery.rtreeQuery              300000  avgt   20   110.227 ± 19.660  us/op
BenchmarkJtsStrTreeQuery.rtreeQuery             1000000  avgt   20   458.004 ± 25.899  us/op

# Improved (degree 16)
Benchmark                           (numBuildEnvelopes)  Mode  Cnt     Score    Error  Units
BenchmarkFlatbushQuery.rtreeQuery                  1000  avgt   20     0.823 ±  0.029  us/op
BenchmarkFlatbushQuery.rtreeQuery                  3000  avgt   20     1.249 ±  0.034  us/op
BenchmarkFlatbushQuery.rtreeQuery                 10000  avgt   20     2.561 ±  0.104  us/op
BenchmarkFlatbushQuery.rtreeQuery                 30000  avgt   20     5.265 ±  0.222  us/op
BenchmarkFlatbushQuery.rtreeQuery                100000  avgt   20    14.124 ±  0.612  us/op
BenchmarkFlatbushQuery.rtreeQuery                300000  avgt   20    58.914 ±  1.750  us/op
BenchmarkFlatbushQuery.rtreeQuery               1000000  avgt   20   238.373 ±  4.453  us/op

# Original
Benchmark                           (numBuildEnvelopes)  Mode  Cnt     Score     Error  Units
BenchmarkJtsStrTreeBuild.buildRtree                1000  avgt   20     0.377 ±   0.009  ms/op
BenchmarkJtsStrTreeBuild.buildRtree                3000  avgt   20     1.368 ±   0.037  ms/op
BenchmarkJtsStrTreeBuild.buildRtree               10000  avgt   20     5.495 ±   0.108  ms/op
BenchmarkJtsStrTreeBuild.buildRtree               30000  avgt   20    19.546 ±   0.383  ms/op
BenchmarkJtsStrTreeBuild.buildRtree              100000  avgt   20    81.362 ±   2.989  ms/op
BenchmarkJtsStrTreeBuild.buildRtree              300000  avgt   20   357.022 ±   4.268  ms/op
BenchmarkJtsStrTreeBuild.buildRtree             1000000  avgt   20  1574.333 ±  48.949  ms/op

# Improved (degree 16)
Benchmark                           (numBuildEnvelopes)  Mode  Cnt     Score    Error  Units
BenchmarkFlatbushBuild.buildRtree                  1000  avgt   20     0.647 ±  0.015  ms/op
BenchmarkFlatbushBuild.buildRtree                  3000  avgt   20     2.164 ±  0.045  ms/op
BenchmarkFlatbushBuild.buildRtree                 10000  avgt   20     4.053 ±  0.152  ms/op
BenchmarkFlatbushBuild.buildRtree                 30000  avgt   20    12.564 ±  0.138  ms/op
BenchmarkFlatbushBuild.buildRtree                100000  avgt   20    32.279 ±  0.190  ms/op
BenchmarkFlatbushBuild.buildRtree                300000  avgt   20   109.482 ±  1.089  ms/op
BenchmarkFlatbushBuild.buildRtree              1000000  avgt    20   435.341 ±  4.130  ms/op