@JunhyungSong 2022-01-18
The test failure(io.trino.plugin.hive.TestHiveFailureRecoveryFile.testJoin) does not seem related to this PR.