@raunaqmorarka 2022-03-17
@JunhyungSong I plan to take a look at it, but it'll take me a while to provide any useful comments as the geospatial features are completely unknown to me. Meanwhile, it'll help if you can rebase this to latest master and organise this PR into smaller commits (perhaps we can start with the commits from prestodb as-is and squash some of the bug fixes if that makes sense).